requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.6.2.post1
audioread==3.0.1
blinker==1.9.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
coqui-tts==0.24.2
coqui-tts-trainer==0.1.4
cutlet==0.4.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
eval_type_backport==0.2.0
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
fugashi==1.3.2
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
gradio==5.1.0
gradio_client==1.4.0
grpcio==1.67.1
grpcio-status==1.67.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
httpcore==1.0.6
httplib2==0.22.0
huggingface-hub==0.25.2
idna==3.10
inflect==7.4.0
jaconv==0.4.0
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
marisa-trie==1.2.1
Markdown==3.7
matplotlib==3.9.2
mojimoji==0.0.13
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
narwhals==1.13.3
networkx==3.4.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
orjson==3.10.7
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.2
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyparsing==3.2.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
regex==2024.9.11
rich==13.9.2
rsa==4.9
ruff==0.7.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
smart-open==7.0.5
smmap==5.0.1
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.40.0
streamlit==1.40.0
streamlit-audiorec==0.1.3
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
together==1.3.3
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
transformers==4.42.4
triton==3.0.0
typeguard==4.3.0
typer==0.12.5
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
watchdog==5.0.3
weasel==0.4.1
websockets==12.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.15.2
